+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.1.1] - 2019-11-15
+### Fixed
+- Don't register Domain denormalizer if Symfony serializer isn't register
+
+## [2.1.0] - 2019-11-12
+### Added
+- Domain component now shows events in the profiler of Symfony!
+
 ## [2.0.3] - 2019-07-25
 ### Changed
 - Removed BC layer for Sf < 4.3 (was useless)